Revert "Closed #16"
This reverts commit e507b3eee7ed0cc39680c2b479531fd51822e94b.

+++ IVS/IVS/math.cs
@@ -118,7 +118,7 @@ namespace MathLibrary
 		/// <param name="x">Zaklad</param>
 		/// <param name="n">N-ta odmocnina</param>
 		/// <returns>Vraci odmocninu s presnosti mensi jak 10^-10</returns>
-		public static double Odmocnina(double x, double n)
+		public static double Odmocnina(double x, int n)
 		{
 			if (Test_Int(n) == false || n < 0 || x < 0)
 			{
